pyproject.toml
[project]
name = "streamlit-supabase-app"
version = "1.0.0"
description = "Aplicación web de gestión de tareas con Streamlit y Supabase"
authors = [
    {name = "Equipo de Desarrollo", email = "dev@example.com"}
]
dependencies = [
    "streamlit>=1.28.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "plotly>=5.15.0",
    "ruff>=0.1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0"
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.ruff.format]
quote-style = "double"